Update json parsing
